ticket_system/requirements.txt
black==25.9.0
python-dotenv==1.1.1
django==5.2.6
djangorestframework==3.16.1
django-filter==25.1